pyproject.toml
[project]
name = "ecomweb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "supabase>=2.15.1",
    "uvicorn>=0.34.2",
]
